LayoutTests/webcl/tests/webcl-test-context-properties.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Testing WebCLContextProperties</title>
<script src="../resources/js-test-pre.js"></script>
<script src="../resources/kernels.js"></script>

<script>

var cl;
var contextProperties;
var context;
var devices;
var deviceType;
var platform;

function runTest() {

    try {
        cl = webcl;
        if (!cl) {
            testFailed("WebCL is not available.");
            return;
        }

        context = cl.createContext(cl.DEVICE_TYPE_DEFAULT);
        if (!context) {
            testFailed("Failed to create WebCLContext");
            return;
        }

        contextDevices = context.getInfo(cl.CONTEXT_DEVICES);
        contextDevicesCount = context.getInfo(cl.CONTEXT_NUM_DEVICES);
        devices = contextDevices;
        shouldBeTrue("devices.length === contextDevicesCount");
        shouldBeTrue("devices[0] instanceof WebCLDevice");

    } catch (e) {
        testFailed("Test WebCLContextProperties has failed; Message: " + e.message);
    }
}

</script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
    description(document.title);
    runTest();
</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
